+++ CHANGELOG.md
@@ -5,12 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 2024-08-02
+
+### Added
+
+- Added Rector.
+
+### Changed
+
+- Use PHPUnit instead of Pest.
+
 ## 2024-07-31
 
 ### Added
 
 - The ability to (also) remember Bible verses.
-- The ability to remove all (learnt) lessons and verses.
+- The ability to remove all (learnt) lessons and verses.bash
 - Production build via FrankenPHP.
 
 ### Changed